The Settlers 2 — Update script to 2.11 syntax

+++ play.it-2/games/play-the-settlers-2.sh
@@ -1,8 +1,8 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
-# Copyright (c) 2015-2018, Antoine Le Gonidec
+# Copyright (c) 2015-2019, Antoine Le Gonidec
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,11 +30,11 @@ set -o errexit
 
 ###
 # The Settlers 2
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20180303.1
 
 # Set game-specific variables
 
@@ -62,19 +62,19 @@ ARCHIVE_GOG_DE_VERSION='1.51-gog2.1.0.17'
 ARCHIVE_GOG_DE_SIZE='370000'
 
 ARCHIVE_DOC_DATA_PATH='app'
-ARCHIVE_DOC_DATA_FILES='./eula ./*.txt'
+ARCHIVE_DOC_DATA_FILES='eula *.txt'
 
 ARCHIVE_GAME_BIN_PATH='app'
-ARCHIVE_GAME_BIN_FILES='./dos4gw.exe ./s2edit.exe ./s2.exe ./setup.exe ./setup.ini video/smackply.exe ./drivers'
+ARCHIVE_GAME_BIN_FILES='dos4gw.exe s2edit.exe s2.exe setup.exe setup.ini video/smackply.exe drivers'
 
 ARCHIVE_GAME_DATA_PATH='app'
-ARCHIVE_GAME_DATA_FILES='./settlers2.gog ./settlers2.inst ./settler2.vmc ./data/animdat ./data/bobs ./data/*.lst ./data/cbob ./data/*.dat ./data/*.idx ./data/io/*.dat ./data/io/*.fnt ./data/masks ./data/mbob ./data/missions/mis_00*.rtx ./data/missions/mis_10*.rtx ./data/sounddat ./data/textures ./gfx/palette ./gfx/pics2 ./gfx/pics/install.lbm ./gfx/pics/mission ./gfx/pics/setup013.lbm ./gfx/pics/setup015.lbm ./gfx/pics/setup666.lbm ./gfx/pics/setup667.lbm ./gfx/pics/setup801.lbm ./gfx/pics/setup802.lbm ./gfx/pics/setup803.lbm ./gfx/pics/setup804.lbm ./gfx/pics/setup805.lbm ./gfx/pics/setup806.lbm ./gfx/pics/setup810.lbm ./gfx/pics/setup811.lbm ./gfx/pics/setup895.lbm ./gfx/pics/setup896.lbm ./gfx/pics/setup899.lbm ./gfx/pics/setup990.lbm ./gfx/pics/world.lbm ./gfx/pics/worldmsk.lbm ./gfx/textures ./gfw_high.ico ./goggame-1207658786.ico'
+ARCHIVE_GAME_DATA_FILES='settlers2.gog settlers2.inst settler2.vmc data/animdat data/bobs data/*.lst data/cbob data/*.dat data/*.idx data/io/*.dat data/io/*.fnt data/masks data/mbob data/missions/mis_00*.rtx data/missions/mis_10*.rtx data/sounddat data/textures gfx/palette gfx/pics2 gfx/pics/install.lbm gfx/pics/mission gfx/pics/setup013.lbm gfx/pics/setup015.lbm gfx/pics/setup666.lbm gfx/pics/setup667.lbm gfx/pics/setup801.lbm gfx/pics/setup802.lbm gfx/pics/setup803.lbm gfx/pics/setup804.lbm gfx/pics/setup805.lbm gfx/pics/setup806.lbm gfx/pics/setup810.lbm gfx/pics/setup811.lbm gfx/pics/setup895.lbm gfx/pics/setup896.lbm gfx/pics/setup899.lbm gfx/pics/setup990.lbm gfx/pics/world.lbm gfx/pics/worldmsk.lbm gfx/textures gfw_high.ico goggame-1207658786.ico'
 
-ARCHIVE_GAME1_L10N_PATH='app'
-ARCHIVE_GAME1_L10N_FILES='./install.scr ./video/intro.smk ./save/mission.dat ./gfx/pics/setup000.lbm ./gfx/pics/setup010.lbm ./gfx/pics/setup011.lbm ./gfx/pics/setup012.lbm ./gfx/pics/setup014.lbm ./gfx/pics/setup897.lbm ./gfx/pics/setup898.lbm ./gfx/pics/setup900.lbm ./gfx/pics/setup901.lbm ./gfx/pics/setup996.lbm ./gfx/pics/setup997.lbm ./gfx/pics/setup998.lbm ./data/resource.idx ./data/missions/mis_0100.rtx ./data/txt* ./data/online ./data/maps* ./data/io/editio.idx ./data/io/io.idx'
+ARCHIVE_GAME0_L10N_PATH='app'
+ARCHIVE_GAME0_L10N_FILES='install.scr video/intro.smk save/mission.dat gfx/pics/setup000.lbm gfx/pics/setup010.lbm gfx/pics/setup011.lbm gfx/pics/setup012.lbm gfx/pics/setup014.lbm gfx/pics/setup897.lbm gfx/pics/setup898.lbm gfx/pics/setup900.lbm gfx/pics/setup901.lbm gfx/pics/setup996.lbm gfx/pics/setup997.lbm gfx/pics/setup998.lbm data/resource.idx data/missions/mis_0100.rtx data/txt* data/online data/maps* data/io/editio.idx data/io/io.idx'
 
-ARCHIVE_GAME2_L10N_PATH='app/__support/save'
-ARCHIVE_GAME2_L10N_FILES='save/mission.dat'
+ARCHIVE_GAME1_L10N_PATH='app/__support/save'
+ARCHIVE_GAME1_L10N_FILES='save/mission.dat'
 
 CONFIG_FILES='./setup.ini'
 DATA_DIRS='./data ./gfx ./save ./worlds'
@@ -88,7 +88,6 @@ APP_MAIN_PRERUN='@video\smackply video\intro.smk'
 APP_MAIN_ICON_GOG_EN='gfw_high.ico'
 APP_MAIN_ICON_GOG_FR='goggame-1207658786.ico'
 APP_MAIN_ICON_GOG_DE='goggame-1207658786.ico'
-APP_MAIN_ICON_RES='16 32 48 256'
 
 APP_EDITOR_TYPE='dosbox'
 APP_EDITOR_ID="${GAME_ID}_edit"
@@ -128,46 +127,48 @@ PKG_BIN_DEPS="$PKG_DATA_ID $PKG_L10N_ID dosbox"
 
 # Load common functions
 
-target_version='2.3'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
-
-for PKG in $PACKAGES_LIST; do
-	organize_data "DOC_${PKG#PKG_}"   "$PATH_DOC"
-	organize_data "GAME_${PKG#PKG_}"  "$PATH_GAME"
-	organize_data "GAME1_${PKG#PKG_}" "$PATH_GAME"
-	organize_data "GAME2_${PKG#PKG_}" "$PATH_GAME"
-done
-
+prepare_package_layout
 sed --in-place 's/SETTLERS2.gog/settlers2.gog/' "${PKG_DATA_PATH}${PATH_GAME}/$GAME_IMAGE"
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
+
+# Extract icons
 
 PKG='PKG_DATA'
-extract_and_sort_icons_from 'APP_MAIN'
-rm "${PKG_DATA_PATH}${PATH_GAME}/$APP_MAIN_ICON"
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN' 'APP_EDITOR' 'APP_SETUP'
+launchers_write 'APP_MAIN' 'APP_EDITOR' 'APP_SETUP'
 
 # Build package
